feat(repl): implement staged runbook with anti-hallucination guarantees (054)
- Add migration 054: staged_runbook, staged_command, staged_command_entity, staged_command_candidate tables
- Implement rust/src/repl/ module:
  - staged_runbook.rs: Rust types (StagedRunbook, StagedCommand, ResolutionStatus)
  - repository.rs: DB access layer
  - resolver.rs: EntityArgResolver for shorthand → UUID resolution
  - dag_analyzer.rs: Dependency detection and topological sort
  - events.rs: MCP event types (CommandStaged, ResolutionAmbiguous, etc.)
  - service.rs: RunbookService with stage/pick/run/show/preview/remove/abort
- Add MCP handlers (runbook.rs) and register 7 runbook tools
- Add runbook.yaml verb definitions and runbook_ops.rs CustomOperations
- Add integration tests (staged_runbook_integration.rs)

Key guarantees:
- No side-effects without explicit run/execute/commit
- No invented UUIDs - all from DB resolution or picker validation
- Server-side ready gate rejects run if not all resolved
- DAG ordering with transparent reorder diff

Also: clippy fixes, CLAUDE.md update, schema_export.sql refresh

+++ rust/src/domain_ops/client_group_ops.rs
@@ -47,6 +47,7 @@ use sqlx::PgPool;
 // RESULT TYPES
 // =============================================================================
 
+#[allow(dead_code)] // Future use for membership operations
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EntityMembershipResult {
     pub id: Uuid,
@@ -723,7 +724,7 @@ impl CustomOperation for ClientGroupSearchOp {
                 entity_id: r.entity_id,
                 entity_name: r.entity_name,
                 matched_tag: r.tag,
-                confidence: r.confidence as f64,
+                confidence: r.confidence,
                 match_type: r.match_type,
             })
             .collect();
